auth and passport implemented

+++ TrashReminderApi/app/Http/Controllers/GarbageController.php
@@ -15,7 +15,7 @@ class GarbageController extends Controller
      */
     public function index()
     {
-        return Garbage::all();
+        return auth()->user()->UserGarbage();
     }
 
     /**
@@ -51,6 +51,7 @@ class GarbageController extends Controller
             $garbage->ora_inizio = $request->ora_inizio;
             $garbage->ora_fine = $request->ora_fine;
             $garbage->giorno = $request->giorno;
+            $garbage->user_id = auth()->user()->id;
             $result = $garbage->save();
             if ($result) {
                 return ["Result" => "Data has been saved"];
@@ -68,7 +69,13 @@ class GarbageController extends Controller
      */
     public function show($id)
     {
-        return Garbage::find($id);
+        $result = auth()->user()->UserGarbage()->find($id);
+        if($result)
+        {
+            return $result;
+        }else{
+            return response('Not found.', 404);
+        }
     }
 
     /**
@@ -93,25 +100,29 @@ class GarbageController extends Controller
     /////////////////////////////////////////////////
     public function update(Request $request, $id)
     {
-        $rules = array(
-            "tipo" => "required",
-            "ora_inizio" => "required|min:5|max:5",
-            "ora_fine" => "required|min:5|max:5",
-            "giorno" => "required|min:6|max:9",
-        );
-        $validator = Validator::make($request->all(), $rules);
-        if ($validator->fails()) {
-            return response()->json($validator->errors(), 401);
-        } else {
-            $garbage = new Garbage;
-            $garbage->name = $request->name;
-            $garbage->description = $request->description;
-            $result = $garbage->save();
-            if ($result) {
-                return ["result" => "data has been created"];
+        $garbage = Garbage::find($id);
+        if (auth()->user()->id == $garbage->user_id) {
+            $rules = array(
+                "tipo" => "required",
+                "ora_inizio" => "required|min:5|max:5",
+                "ora_fine" => "required|min:5|max:5",
+                "giorno" => "required|min:6|max:9",
+            );
+            $validator = Validator::make($request->all(), $rules);
+            if ($validator->fails()) {
+                return response()->json($validator->errors(), 401);
             } else {
-                return ["result" => "operation failed"];
+                $garbage->name = $request->name;
+                $garbage->description = $request->description;
+                $result = $garbage->save();
+                if ($result) {
+                    return ["result" => "data has been created"];
+                } else {
+                    return ["result" => "operation failed"];
+                }
             }
+        } else {
+            return response('Unauthenticated.', 401);
         }
     }
 
@@ -124,11 +135,15 @@ class GarbageController extends Controller
     public function destroy($id)
     {
         $garbage = Garbage::find($id);
-        $result = $garbage->delete();
-        if ($result) {
-            return ["result" => "data has been deleted"];
+        if (auth()->user()->id == $garbage->user_id) {
+            $result = $garbage->delete();
+            if ($result) {
+                return ["result" => "data has been deleted"];
+            } else {
+                return ["result" => "operation failed"];
+            }
         } else {
-            return ["result" => "operation failed"];
+            return response('Unauthenticated.', 401);
         }
     }
 }
